+++ pages/infoAdmins/infoAdm.tsx
@@ -2,11 +2,12 @@ import * as React from 'react';
 import useSWR from 'swr'
 import { DataGrid, GridColDef, GridValueGetterParams } from '@mui/x-data-grid';
 import { getMaxListeners } from 'process';
-import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
+
 import Button from '@mui/material/Button';
 import estilos from '../infoAdmins/info.module.css'
 import AddIcon from '../infoAdmins/modalAdd'
 import { ClassNames } from '@emotion/react';
+
 
 
 const columns: GridColDef[] = [
@@ -58,7 +59,6 @@ export default function DataTable() {
       
     </div>
     <AddIcon/>
-
     </React.Fragment>
     
   );
